requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
alabaster==1.0.0
arrow==1.3.0
astroid==3.3.9
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
asyncssh==2.20.0
atomicwrites==1.4.1
attrs==25.3.0
autopep8==2.0.4
babel==2.17.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
binaryornot==0.4.4
black==25.1.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cookiecutter==2.6.0
cryptography==44.0.2
cycler==0.12.1
deap==1.4.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
diff-match-patch==20241021
dill==0.3.9
docstring-to-markdown==0.15
docutils==0.21.2
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
flake8==7.1.2
Flask==3.1.0
Flask-JWT-Extended==4.7.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
gast==0.6.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.6.1
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.34.0
isort==6.0.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jellyfish==1.1.3
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==2.15.0
keyring==25.6.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.2
ml-dtypes==0.2.0
more-itertools==10.6.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.23.5
numpydoc==1.8.0
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydocstyle==6.3.0
pyflakes==3.2.0
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.5
pylint-venv==3.0.4
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2025.1
pyuca==1.2
pywin32==309
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.3.0
QDarkStyle==3.2.3
qstylizer==0.2.4
QtAwesome==1.3.1
qtconsole==5.6.1
QtPy==2.4.3
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rope==1.13.0
rpds-py==0.23.1
rsa==4.9
rtree==1.4.0
scikeras==0.12.0
scikit-learn==1.3.0
scipy==1.10.0
seaborn==0.13.2
six==1.17.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spyder==6.0.4
spyder-kernels==3.0.3
stack-data==0.6.3
superqt==0.7.1
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
text-unidecode==1.3
textdistance==4.6.3
threadpoolctl==3.6.0
three-merge==0.1.1
tinycss2==1.4.0
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
whatthepatch==1.0.7
wrapt==1.14.1
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
